requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-cors-headers==2.1.0
django-sslserver==0.19
djangorestframework==3.7.1
idna==2.6
pytz==2017.3
requests==2.18.4
urllib3==1.22
whitenoise==3.2
